requirements.txt
xarray
numpy
h5netcdf
scipy
dask
zarr
hydra-core
torch